Add useBreedingEggs composable
