+++ src/auto-imports.d.ts
@@ -172,6 +172,7 @@ declare global {
   const useBattleship: typeof import('./composables/useBattleship')['useBattleship']
   const useBluetooth: typeof import('@vueuse/core')['useBluetooth']
   const useBreakpoints: typeof import('@vueuse/core')['useBreakpoints']
+  const useBreedingEggs: typeof import('./composables/useBreedingEggs')['useBreedingEggs']
   const useBreedingStore: typeof import('./stores/breeding')['useBreedingStore']
   const useBroadcastChannel: typeof import('@vueuse/core')['useBroadcastChannel']
   const useBrowserLocation: typeof import('@vueuse/core')['useBrowserLocation']
@@ -430,6 +431,9 @@ declare global {
   export type { Cell } from './composables/useBattleship'
   import('./composables/useBattleship')
   // @ts-ignore
+  export type { BreedingEntry } from './composables/useBreedingEggs'
+  import('./composables/useBreedingEggs')
+  // @ts-ignore
   export type { FaintAutoEmitResult } from './composables/useFaintAutoEmit'
   import('./composables/useFaintAutoEmit')
   // @ts-ignore
@@ -674,6 +678,7 @@ declare module 'vue' {
     readonly useBattleship: UnwrapRef<typeof import('./composables/useBattleship')['useBattleship']>
     readonly useBluetooth: UnwrapRef<typeof import('@vueuse/core')['useBluetooth']>
     readonly useBreakpoints: UnwrapRef<typeof import('@vueuse/core')['useBreakpoints']>
+    readonly useBreedingEggs: UnwrapRef<typeof import('./composables/useBreedingEggs')['useBreedingEggs']>
     readonly useBreedingStore: UnwrapRef<typeof import('./stores/breeding')['useBreedingStore']>
     readonly useBroadcastChannel: UnwrapRef<typeof import('@vueuse/core')['useBroadcastChannel']>
     readonly useBrowserLocation: UnwrapRef<typeof import('@vueuse/core')['useBrowserLocation']>
